streamlit-ui/requirements.txt
streamlit
requests
httpx
python-dotenv
sseclient-py
